[step3] refactor: output 관련한 로직을 값 객체에 위임
